# pb/book.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "book.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "BookService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/books": {
      "get": {
        "operationId": "BookService_ListBooks",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbListBooksResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "BookService"
        ]
      },
      "post": {
        "operationId": "BookService_CreateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateBookRequest"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/v1/books/{data.id}": {
      "put": {
        "operationId": "BookService_UpdateBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbUpdateBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "data.id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BookServiceUpdateBookBody"
            }
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    },
    "/v1/books/{id}": {
      "get": {
        "operationId": "BookService_GetBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbGetBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      },
      "delete": {
        "operationId": "BookService_DeleteBook",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteBookResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BookService"
        ]
      }
    }
  },
  "definitions": {
    "BookServiceUpdateBookBody": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object",
          "properties": {
            "title": {
              "type": "string"
            },
            "author": {
              "type": "string"
            },
            "pages": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      }
    },
    "pbBook": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "pages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "pbCreateBookRequest": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/pbBook"
        }
      }
    },
    "pbCreateBookResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/pbBook"
        }
      }
    },
    "pbDeleteBookResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "pbGetBookResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/pbBook"
        }
      }
    },
    "pbListBooksResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/pbBook"
          }
        }
      }
    },
    "pbUpdateBookResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/pbBook"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}